added tests for update
these should be the last tests before I'm confident this controller works and then I can move on to the next ones before added the rest of the database models (and their repos, controllers, etc). this controller plus tests should be pretty similar to the majority of what I will have.

currently the last two tests are failing. One not implemented and failing on purpose and the other is getting an internal server error :/